+++ packages/svelte-maplibre-storymap/src/lib/index.ts
@@ -1,4 +1,6 @@
 import StoryMap from './StoryMap.svelte';
+import StoryMapChapter from './StoryMapChapter.svelte';
 
-export { StoryMap };
+export { StoryMap, StoryMapChapter };
 export * from './interfaces/index.js';
+export * from './stores/index.js';