+++ packages/svelte-maplibre-storymap/src/lib/interfaces/StoryMapChapter.ts
@@ -1,4 +1,19 @@
-import type { StyleSpecification } from 'maplibre-gl';
+import type { ControlPosition, StyleSpecification } from 'maplibre-gl';
+
+export interface StoryMapChapterLayerEvent {
+	/**
+	 * Layer name as assigned in MapLibre style.
+	 */
+	layer: string;
+	/**
+	 * The opacity to display the layer. 0 is fully transparent, 1 is fully opaque.
+	 */
+	opacity: number;
+	/**
+	 * The length of the opacity transition, numeric, in milliseconds. Default is 300. This is an optional parameter and can be omitted.
+	 */
+	duration?: number;
+}
 
 export interface StoryMapChapter {
 	/**
@@ -37,6 +52,11 @@ export interface StoryMapChapter {
 	mapInteractive?: boolean;
 
 	/**
+	 * Navigation control's position. Default is 'top-right'
+	 */
+	mapNavigationPosition?: ControlPosition;
+
+	/**
 	 * Details about the map display and camera view.
 	 */
 	location: {
@@ -72,4 +92,14 @@ export interface StoryMapChapter {
 	 * Use different style for the chapter from main one if it is set.
 	 */
 	style?: StyleSpecification | string;
+
+	/**
+	 * Layers to be displayed/hidden/muted when the section becomes active. Array of objects
+	 */
+	onChapterEnter?: StoryMapChapterLayerEvent[];
+
+	/**
+	 * Same as onChapterEnter except it is triggered when the section becomes inactive. Array of objects
+	 */
+	onChapterExit?: StoryMapChapterLayerEvent[];
 }