+++ sites/geohub/src/components/util/MiniMap.svelte
@@ -18,6 +18,7 @@
 	const dispatch = createEventDispatcher();
 
 	export let feature: DatasetFeature;
+
 	export let width = '100%';
 	export let height = '100%';
 	export let isLoadMap = false;
@@ -34,7 +35,6 @@
 	export let metadata: RasterTileMetadata | VectorTileMetadata = undefined;
 	const is_raster: boolean = feature.properties.is_raster as unknown as boolean;
 	const url: string = feature.properties.url;
-
 	let rasterTile: RasterTileData;
 	let vectorTile: VectorTileData;