+++ sites/geohub/src/routes/(app)/+layout@.svelte
@@ -33,5 +33,5 @@
 <style global lang="scss">
 	@import 'bulma-switch/dist/css/bulma-switch.min.css';
 	@import 'bulma-divider/dist/css/bulma-divider.min.css';
-	@import '/node_modules/flag-icons/css/flag-icons.min.css';
+	@import 'flag-icons/css/flag-icons.min.css';
 </style>