+++ sites/geohub/src/routes/api/stac/style/[id]/[collection]/[...item]/[asset]/+server.ts
@@ -2,6 +2,8 @@ import { createDatasetLinks } from '$lib/server/helpers';
 import type { RequestHandler } from './$types';
 import RasterDefaultStyle from '$lib/server/defaultStyle/RasterDefaultStyle';
 import type { UserConfig } from '$lib/config/DefaultUserConfig';
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-expect-error
 import { env } from '$env/dynamic/private';
 
 export const GET: RequestHandler = async ({ url, fetch }) => {