+++ packages/svelte-undp-components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @undp-data/svelte-undp-components
 
+## 1.0.2
+
+### Patch Changes
+
+- 685729f: fix: applied SohneBreit and other formats of UNDP design for HeroHeader component.
+
+## 1.0.1
+
+### Patch Changes
+
+- 0c7d0a8: fix: add breadcrumbClicked event in HeroHeader component
+
 ## 1.0.0
 
 ### Major Changes