+++ packages/undp-bulma/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @undp-data/undp-bulma
 
+## 1.0.1
+
+### Patch Changes
+
+- 685729f: fix: applied UNDP design system style for all bulma headings.
+
 ## 1.0.0
 
 ### Major Changes